% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeftestCluster.R
\name{coeftestCluster}
\alias{coeftestCluster}
\title{Cluster standard errors on one or two variables}
\usage{
coeftestCluster(data, mod, namevec, cluster1, cluster2 = NULL, coeftest = T)
}
\arguments{
\item{data}{Dataframe used to estimate model}

\item{mod}{Model estimated using \code{glm} or \code{lm}}

\item{namevec}{Names of model variables}

\item{cluster1}{Clustering variable}

\item{cluster2}{Second clustering variable (optional)}

\item{coeftest}{Boolean. Return \code{coeftest} object (TRUE) or clustered variance-covariance matrix?}
}
\value{
\code{coeftest} object or a clustered variance-covariance matrix
}
\description{
Slightly modified function to cluster standard errors from \code{glm} or \code{lm} models.
Taken from: \url{https://github.com/iangow-public/acct_data/blob/master/code/cluster2.R}
}
