% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfAsList.R
\name{dataAsList}
\alias{dataAsList}
\title{Function to convert dataframe into list organized by grouping variable}
\usage{
dataAsList(data, groupingvar)
}
\arguments{
\item{data}{Dataframe}

\item{groupingvar}{grouping variable}
}
\value{
List with slots defined by grouping variable
}
\description{
Function to convert dataframe into list organized by grouping variable
}
