% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSeries.R
\name{generateSeries}
\alias{generateSeries}
\title{Expand interval dataframe into series}
\usage{
generateSeries(
  data,
  start,
  stop,
  step = 1,
  timeint = "day",
  varname = "generate_series"
)
}
\arguments{
\item{data}{Dataframe}

\item{start}{Variable that defines starting point of interval / series}

\item{stop}{Variable that defines end point of interval / series}

\item{step}{Step size of output series. Default: 1}

\item{timeint}{Time increment, for dates only; Days, weeks, months, quarters or years. Default: days}

\item{varname}{Name of series variable. Default: "generate_series"}
}
\value{
Dataframe containing series
}
\description{
Expand dataframe containing an interval into series defined by step size
}
