% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGroupedNA.R
\name{plotGroupedNA}
\alias{plotGroupedNA}
\title{Plot missing values grouped by a variable}
\usage{
plotGroupedNA(data, groupvar)
}
\arguments{
\item{data}{Dataframe}

\item{groupvar}{Variable to group missing observations by}
}
\value{
\code{ggplot2} object
}
\description{
Returns a tile plot with percentages of missing observations for each grouped variable
}
