% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNA.R
\name{plotNA}
\alias{plotNA}
\title{Plot missing values}
\usage{
plotNA(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
\code{ggplot2} object
}
\description{
Returns a bar plot with missing values for each variable
}
