% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_GS.R
\name{theme_GS}
\alias{theme_GS}
\title{Custom ggplot theme using minimal elements}
\usage{
theme_GS()
}
\description{
Theme uses white background, minimal necessary elements,
black axis text and slightly larger font sizes for axis text and titles
than the default.
}
